Update conditional test critical values in light of revisions in ARP (2021)
In the latest revision to Andrews, Roth & Pakes (2021) (https://scholar.harvard.edu/files/iandrews/files/conditonal_linear_moments_r1-_main.pdf), the conditional test's critical value is now defined as max(0, c_{1-alpha)}, where c_{1-alpha} is the 1-alpha quantile of an appropriately defined truncated normal distribution.

Andrews, Roth & Pakes (2021) redefine the conditional critical value in this way in order to simplify uniform asymptotic arguments, but it is not necessary for size control in the finite sample normal model. Furthermore, it enforces that the test does not reject when all the moment inequalities are satisfied in sample.

In this commit, we update the calculation of the conditional critical value to reflect this truncation at zero in the non-nuisance parameter functions. Prior versions of the package implemented the conditional test as described in the previous draft Andrews, Roth & Pakes (2019), where the conditional critical value was simply defined as c_{1-alpha}.

In arp-nonuisance.R, this involves changes to the computation of the conditional critical value in
- .testInIdentifiedSet()
- .testInIdentifiedSet_LF_Hybrid()
This commit also corrects a bug in the computation of the modified size for the least-favorable hybrid test without nuisance parameters.

In arp-nuisance.R, this involves changes to the computation of the conditional critical value in
- .lp_conditional_test_fn()