Improved performance of .vlo_vup_dual_fn by ~order of magnitude
In practice the bisection takes a few dozen iterations, which is the
main bottleneck of the function as far as I have been able to profile.
We want to find c s.t. max_x {a x + b x c} = c. We can instead iterate
by setting c' = - a x / (1 - b x). In practice this  takes one or two
iterations and speeds up computations by an order of magnitude (I think
it is also more precise, since there is only one tolerance involved;
with the bisection the tolerance for equality cannot be set too low).